Extract simulateDelegatecall to compat handler
