+++ contracts/common/StorageAccessible.sol
@@ -4,9 +4,6 @@ pragma solidity >=0.7.0 <0.9.0;
 /// @title StorageAccessible - generic base contract that allows callers to access all internal storage.
 /// @notice Adjusted version of https://github.com/gnosis/util-contracts/blob/3db1e531cb243a48ea91c60a800d537c1000612a/contracts/StorageAccessible.sol
 contract StorageAccessible {
-    // bytes4(keccak256("simulateDelegatecallInternal(address,bytes)"));
-    bytes4 internal constant SIMULATE_DELEGATECALL_INTERNAL_SELECTOR = 0x43218e19;
-
     /**
      * @dev Reads `length` bytes of storage in the currents contract
      * @param offset - the offset in the current contract's storage in words to start reading from
@@ -30,57 +27,22 @@ contract StorageAccessible {
 
     /**
      * @dev Performs a delegetecall on a targetContract in the context of self.
-     * Internally reverts execution to avoid side effects (making it static). Catches revert and returns encoded result as bytes.
-     * @param targetContract Address of the contract containing the code to execute.
-     * @param calldataPayload Calldata that should be sent to the target contract (encoded method name and arguments).
-     */
-    function simulateDelegatecall(
-        address targetContract,
-        bytes memory calldataPayload
-    ) public returns (bytes memory) {
-        bytes memory innerCall = abi.encodeWithSelector(
-            SIMULATE_DELEGATECALL_INTERNAL_SELECTOR,
-            targetContract,
-            calldataPayload
-        );
-        (, bytes memory response) = address(this).call(innerCall);
-        bool innerSuccess = response[response.length - 1] == 0x01;
-        setLength(response, response.length - 1);
-        if (innerSuccess) {
-            return response;
-        } else {
-            revertWith(response);
-        }
-    }
-
-    /**
-     * @dev Performs a delegetecall on a targetContract in the context of self.
      * Internally reverts execution to avoid side effects (making it static). Returns encoded result as revert message
      * concatenated with the success flag of the inner call as a last byte.
      * @param targetContract Address of the contract containing the code to execute.
      * @param calldataPayload Calldata that should be sent to the target contract (encoded method name and arguments).
      */
-    function simulateDelegatecallInternal(
+    function simulate(
         address targetContract,
-        bytes memory calldataPayload
-    ) public returns (bytes memory) {
+        bytes calldata calldataPayload
+    ) external returns (bytes memory) {
         (bool success, bytes memory response) = targetContract.delegatecall(
             calldataPayload
         );
-        revertWith(abi.encodePacked(response, success));
-    }
-
-    function revertWith(bytes memory response) internal pure {
+        bytes memory responseWithStatus = abi.encodePacked(response, success);
         // solium-disable-next-line security/no-inline-assembly
         assembly {
-            revert(add(response, 0x20), mload(response))
-        }
-    }
-
-    function setLength(bytes memory buffer, uint256 length) internal pure {
-        // solium-disable-next-line security/no-inline-assembly
-        assembly {
-            mstore(buffer, length)
+            revert(add(responseWithStatus, 0x20), mload(responseWithStatus))
         }
     }
 }
