+++ test/core/GnosisSafe.StorageAccessible.spec.ts
@@ -1,8 +1,9 @@
 import { expect } from "chai";
-import hre, { deployments, waffle } from "hardhat";
+import { deployments, waffle } from "hardhat";
 import "@nomiclabs/hardhat-ethers";
-import { deployContract, getSafeSingleton, getDefaultCallbackHandler, getSafeWithOwners } from "../utils/setup";
-import { BigNumber, utils } from "ethers";
+import { getSafeSingleton, getDefaultCallbackHandler, getSafeWithOwners } from "../utils/setup";
+import { utils } from "ethers";
+import { killLibContract } from "../utils/contracts";
 
 describe("StorageAccessible", async () => {
 
@@ -11,38 +12,7 @@ describe("StorageAccessible", async () => {
     const setupTests = deployments.createFixture(async ({ deployments }) => {
         await deployments.fixture();
         const handler = await getDefaultCallbackHandler()
-        const source = `
-            contract Test {
-                function killme() public {
-                    selfdestruct(payable(msg.sender));
-                }
-
-                function expose() public returns (address handler) {
-                    bytes32 slot = 0x6c9a6c4a39284e37ed1cf53d337577d14212a4870fb976a4366c693b939918d5;
-                    assembly {
-                        handler := sload(slot)
-                    }
-                }
-
-                function estimate(address to, bytes memory data) public returns (uint256) {
-                    uint256 startGas = gasleft();
-                    (bool success,) = to.call{ gas: gasleft() }(data);
-                    require(success, "Transaction failed");
-                    return startGas - gasleft();
-                }
-
-                address singleton;
-                uint256 public value = 0;
-                function updateAndGet() public returns (uint256) {
-                    value++;
-                    return value;
-                }
-
-                function trever() public returns (address handler) {
-                    revert("Why are you doing this?");
-                }
-            }`
-        const killLib = await deployContract(user1, source);
+        const killLib = await killLibContract(user1);
         return {
             safe: await getSafeWithOwners([user1.address, user2.address], 1, handler.address),
             killLib,
@@ -70,73 +40,26 @@ describe("StorageAccessible", async () => {
         })
     })
 
-    describe("simulateDelegatecall", async () => {
-
-        it('should revert changes', async () => {
-            const { safe, killLib } = await setupTests()
-            const code = await hre.ethers.provider.getCode(safe.address)
-            expect(
-                await safe.callStatic.simulateDelegatecall(killLib.address, killLib.interface.encodeFunctionData("killme"))
-            ).to.be.eq("0x")
-            expect(
-                await hre.ethers.provider.getCode(safe.address)
-            ).to.be.eq(code)
-        })
-
-        it('should return result', async () => {
-            const { safe, killLib, handler } = await setupTests()
-            expect(
-                await safe.callStatic.simulateDelegatecall(killLib.address, killLib.interface.encodeFunctionData("expose"))
-            ).to.be.eq("0x000000000000000000000000" + handler.address.slice(2).toLowerCase())
-        })
-
-        it('should propagate revert message', async () => {
-            const { safe, killLib } = await setupTests()
-            await expect(
-                safe.callStatic.simulateDelegatecall(killLib.address, killLib.interface.encodeFunctionData("trever"))
-            ).to.revertedWith("Why are you doing this?")
-        })
-
-        it('should simulate transaction', async () => {
-            const { safe, killLib } = await setupTests()
-            const estimate = await safe.callStatic.simulateDelegatecall(
-                killLib.address,
-                killLib.interface.encodeFunctionData("estimate", [safe.address, "0x"])
-            )
-            expect(BigNumber.from(estimate).toNumber()).to.be.lte(5000)
-        })
-
-        it('should return modified state', async () => {
-            const { safe, killLib } = await setupTests()
-            const value = await safe.callStatic.simulateDelegatecall(
-                killLib.address,
-                killLib.interface.encodeFunctionData("updateAndGet", [])
-            )
-            expect(BigNumber.from(value).toNumber()).to.be.eq(1)
-            expect((await killLib.value()).toNumber()).to.be.eq(0)
-        })
-    })
-
-    describe("simulateDelegatecallInternal", async () => {
+    describe("simulate", async () => {
 
         it('should revert changes', async () => {
             const { safe, killLib } = await setupTests()
             await expect(
-                safe.callStatic.simulateDelegatecallInternal(killLib.address, killLib.interface.encodeFunctionData("killme"))
+                safe.callStatic.simulate(killLib.address, killLib.interface.encodeFunctionData("killme"))
             ).to.be.reverted
         })
 
         it('should revert the revert with message', async () => {
             const { safe, killLib } = await setupTests()
             await expect(
-                safe.callStatic.simulateDelegatecallInternal(killLib.address, killLib.interface.encodeFunctionData("trever"))
+                safe.callStatic.simulate(killLib.address, killLib.interface.encodeFunctionData("trever"))
             ).to.revertedWith("Why are you doing this?")
         })
 
         it('should return estimate in revert', async () => {
             const { safe, killLib } = await setupTests()
             await expect(
-                safe.callStatic.simulateDelegatecallInternal(killLib.address, killLib.interface.encodeFunctionData("estimate", [safe.address, "0x"]))
+                safe.callStatic.simulate(killLib.address, killLib.interface.encodeFunctionData("estimate", [safe.address, "0x"]))
             ).to.be.reverted
         })
     })
